extract executeTransactionRequest into TransactionRequestExecuting protocol
