+++ Sources/Wallet/Data/Services/Transaction/DefaultTransactionService.swift
@@ -3,25 +3,10 @@ import CrossmintCommonTypes
 import CrossmintService
 import Http
 
-struct DefaultTransactionService: TransactionService, AuthManagerProviding {
+struct DefaultTransactionService: TransactionService, AuthManagerProviding, TransactionRequestExecuting {
     let crossmintService: CrossmintService
     let jsonCoder: JSONCoder
     let authManager: AuthManager
-
-    func executeTransactionRequest<T: WalletTypeTransactionMapping>(
-        endpoint: Endpoint,
-        mapping: T.Type
-    ) async throws(TransactionError) -> any TransactionApiModel {
-        let data = try await crossmintService.executeRequestForRawData(
-            endpoint,
-            errorType: TransactionError.self
-        )
-        do {
-            return try jsonCoder.decode(T.APIModel.self, from: data)
-        } catch {
-            throw TransactionError.transactionGeneric("Failed to decode transaction response")
-        }
-    }
 
     func createTransaction(
         _ request: CreateTransactionRequest
